refactor: 클라이언트 디렉토리 구조 개선 및 레이아웃 분리 - src 계층을 활용하는 쪽으로 디렉토리 구조를 개선했다. - 메인 도메인과 스튜디오 도메인의 레이아웃을 구분하기 위해 도메인에 라우트 그루핑 적용
