+++ client/src/app/(domain)/categories/page.tsx
