+++ client/src/app/(domain)/features/live/NoLiveContent.tsx
