+++ client/src/app/(domain)/features/live/VideoIconButton.tsx
