+++ client/src/app/(domain)/following/page.tsx
