+++ client/src/app/(domain)/lives/[id]/page.tsx
