+++ client/src/app/(domain)/page.tsx
