+++ client/src/app/studio/layout.tsx
@@ -1,0 +1,7 @@
+import { type PropsWithChildren } from 'react';
+
+const layout = ({ children }: PropsWithChildren) => {
+  return <>studio layout{children}</>;
+};
+
+export default layout;
