+++ client/src/hooks/useFocused.ts
