+++ client/src/hooks/useFullscreen.ts
