+++ client/src/hooks/useMouseMovementOnElement.ts
