+++ client/src/libs/internalTypes.ts
