+++ client/src/server/browser.ts
@@ -1,4 +1,4 @@
 import { setupWorker } from 'msw/browser';
-import { handlers } from '@/app/server/handlers';
+import { handlers } from '@server/handlers';
 
 export const worker = setupWorker(...handlers);
