+++ client/src/server/node.ts
@@ -1,4 +1,4 @@
 import { setupServer } from 'msw/node';
-import { handlers } from '@/app/server/handlers';
+import { handlers } from '@server/handlers';
 
 export const server = setupServer(...handlers);
