+++ client/vitest.config.ts
@@ -6,14 +6,14 @@ export default defineConfig({
   plugins: [react()],
   resolve: {
     alias: {
-      '@assets': fileURLToPath(new URL('./src/app/assets', import.meta.url)),
-      '@test': fileURLToPath(new URL('./src/app/__test__', import.meta.url)),
-      '@mocks': fileURLToPath(new URL('./src/app/__mocks__', import.meta.url)),
-      '@hooks': fileURLToPath(new URL('./src/app/hooks', import.meta.url)),
+      '@assets': fileURLToPath(new URL('./src/assets', import.meta.url)),
+      '@test': fileURLToPath(new URL('./src/__test__', import.meta.url)),
+      '@mocks': fileURLToPath(new URL('./src/__mocks__', import.meta.url)),
+      '@hooks': fileURLToPath(new URL('./src/hooks', import.meta.url)),
       '@providers': fileURLToPath(new URL('./src/app/providers', import.meta.url)),
-      '@libs': fileURLToPath(new URL('./src/app/libs', import.meta.url)),
+      '@libs': fileURLToPath(new URL('./src/libs', import.meta.url)),
       '@components': fileURLToPath(new URL('./src/app/components', import.meta.url)),
-      '@server': fileURLToPath(new URL('./src/app/server', import.meta.url)),
+      '@server': fileURLToPath(new URL('./src/server', import.meta.url)),
       '@app': fileURLToPath(new URL('./src/app', import.meta.url)),
       '@': fileURLToPath(new URL('./src', import.meta.url)),
     },
